3619: breakout OrgProfileTab into own dir with subcomponents, and use scss classname universe styling.
